requirements.txt
asgiref==3.7.2
boto3==1.26.39
botocore==1.29.40
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
distlib==0.4.0
dj-database-url==2.1.0
dj-rest-auth==7.0.1
Django==5.0.14
django-allauth==65.0.2
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.7
filelock==3.18.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2==2.9.9
pycountry==23.12.11
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.33.1
whitenoise==6.2.0
